Add tool trait.
